+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `MS3Record.get_extra_header()` to get a specified extra header
 - `MS3Record.set_extra_header()` to set a specified extra header
 - `MS3Record.merge_extra_headers()` to apply a JSON Merge Patch to extra headers
+- `MS3Record.valid_extra_headers()` to validate extra headers
 
 ### Changed
 - libmseed updated to v3.2.3